13：前后端分别实现角色权限控制

+++ springboot/src/main/java/com/example/service/NoticeService.java
@@ -2,8 +2,11 @@ package com.example.service;
 
 import cn.hutool.core.date.DateUnit;
 import cn.hutool.core.date.DateUtil;
+import com.example.entity.Account;
 import com.example.entity.Notice;
+import com.example.exception.CustomerException;
 import com.example.mapper.NoticeMapper;
+import com.example.utils.TokenUtils;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
 import jakarta.annotation.Resource;
@@ -29,15 +32,27 @@ public class NoticeService {
     }
 
     public void add(Notice notice) {
+        Account currentUser = TokenUtils.getCurrentUser();
+        if ("USER".equals(currentUser.getRole())) {
+            throw new CustomerException("500","您现在没有权限进行该操作");
+        }
         notice.setTime(DateUtil.now());
         noticeMapper.insert(notice);
     }
 
     public void update(Notice notice) {
+        Account currentUser = TokenUtils.getCurrentUser();
+        if ("USER".equals(currentUser.getRole())) {
+            throw new CustomerException("500","您现在没有权限进行该操作");
+        }
         noticeMapper.updateById(notice);
     }
 
     public void deleteById(Integer id) {
+        Account currentUser = TokenUtils.getCurrentUser();
+        if ("USER".equals(currentUser.getRole())) {
+            throw new CustomerException("500","您现在没有权限进行该操作");
+        }
         noticeMapper.deleteById(id);
     }
